+++ src/components/SearchResultCard.js
@@ -3,8 +3,7 @@ import React from 'react'
 
 const SearchResultCard = ({ info,theme }) => {
 
-   
-   console.log(info);
+ 
     return (
         <div className=''>
