+++ src/config/category.js
@@ -1,0 +1,15 @@
+const category=[
+    "All",
+    "React",
+    "Songs",
+    "Gaming",
+    "Entertain",
+    "Treaking",
+    "Movies",
+    "Racing",
+    "Bikes",
+    "Engineering",
+    "Riding",
+  ]
+
+export default category
