+++ src/config/constant.js
@@ -2,5 +2,4 @@ export const API_KEY="AIzaSyB8bmu-ptu-aF0mldG3a3JWT5-pccwMS4E"
 export const YOUTUBE_API_URL='https://youtube.googleapis.com/youtube/v3/videos?part=snippet%2CcontentDetails%2Cstatistics&chart=mostPopular&maxResults=24&regionCode=IN&key='+API_KEY+"&pageToken=";
 export const commentsURL="https://youtube.googleapis.com/youtube/v3/commentThreads?part=snippet%2Creplies&&maxResults=20&";
 export const channelDetailURL="https://youtube.googleapis.com/youtube/v3/channels?part=snippet%2CcontentDetails%2Cstatistics&id=";
-
 export const vedioSummary='https://youtube.googleapis.com/youtube/v3/videos?part=snippet%2CcontentDetails%2Cstatistics&id=';
